+++ test/unit/test_arguments.py
@@ -218,7 +218,6 @@ def test_training_arg(test_params, expected_params):
                       length_penalty_beta=0.0,
                       strip_unknown_words=False,
                       override_dtype=None,
-                      skip_softmax=False,
                       skip_topk=False)),
 ])
 def test_inference_args(test_params, expected_params):